requirements.txt
Babel==2.3.4
Django==1.8.13
django-appconf==1.0.2
django-compressor==2.0
django-extra-views==0.6.4
django-haystack==2.4.1
django-oscar==1.2.1
django-tables2==1.0.7
django-treebeard==4.0.1
django-widget-tweaks==1.4.1
factory-boy==2.6.1
fake-factory==0.5.7
mock==1.3.0
pbr==1.9.1
phonenumbers==7.4.0
Pillow==3.2.0
purl==1.3
python-dateutil==2.5.3
pytz==2016.4
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
sorl-thumbnail==12.3
Unidecode==0.4.19
